+++ metadata.rb
@@ -1,7 +1,16 @@
-name             'add-vagrantuser'
+name             'ubuntu-xfce4'
 maintainer       'Lonnie VanZandt'
 maintainer_email 'lonniev@gmail.com'
 license          'Apache 2.0'
-description      'Adds a Vagrant User to a box that lacks one'
+description      'Installs/Configures Ubuntu xfce4'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'
+{"ubuntu" => '12.04', "fedora" => '17.0'}.each do |os,ver|
+  supports os, ">= #{ver}"
+  case os
+    when 'fedora'
+      depends 'yum'
+    when 'ubuntu'
+      depends 'apt'
+  end
+end